consensus/misc: move eip1559 into a package (#27828)
